requirements.txt
# =============================================================================
# Dependencias Python — FinBank Data Pipeline
# =============================================================================

# Generación de datos
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
pyyaml>=6.0

# Carga a SQL Database
sqlalchemy>=2.0.0
pyodbc>=5.0.0  # Azure SQL
# psycopg2-binary>=2.9.0  # PostgreSQL alternativo

# Testing
pytest>=7.0.0

# Calidad de datos (opcional - complemento local)
# great-expectations>=0.18.0
# soda-core>=3.0.0
